+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.16.0] - 2019-10-29
+
 ### Added
 
 - Items can now be added via query string in the format `/cart/add?sku={skuId}&qty={quantity}&seller={seller}`